requirements.txt
tensorflow==1.2.0
Flask==0.12.1
numpy==1.13.0
Pillow==4.1.1
